requirements.txt
openai-whisper
requests==2.32.3
torch
torchaudio
numpy>=1.21.0
dearpygui>=2.0.0
customtkinter
google-genai
sounddevice
pyaudio
packaging
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.6
msal>=1.28.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.127.0